requirements.txt
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
python-dotenv==1.0.1
xgboost==2.0.3
lightgbm==4.3.0
scikit-learn==1.4.2
imbalanced-learn==0.12.2
causalml==0.15.0
shap==0.45.0
optuna==3.6.1
PuLP==2.8.0
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0
fpdf2==2.7.9
reportlab==4.2.0
pytest==8.2.0
black==24.4.2
flake8==7.0.0
isort==5.13.2
pre-commit==3.7.1
PyYAML==6.0.1
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
